+++ yampt.qt/source/translation_example.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <string>
+
+struct translation_example_t
+{
+	std::string original;
+	std::string translation;
+};
+
+constexpr int max_examples = 3;